# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "Documentation for the Car Reservation API",
        "contact": {
            "email": "admin@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/konsultasis": {
            "get": {
                "tags": [
                    "Konsultasi"
                ],
                "summary": "Get all consultations",
                "operationId": "54e552a7fda300a1126cf4c6ea798f51",
                "responses": {
                    "200": {
                        "description": "Konsultasi retrieved successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Konsultasi"
                ],
                "summary": "Create a new consultation",
                "operationId": "3201ed36de5741b6d6fa48bcd9177049",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id_pasien",
                                    "id_dokter",
                                    "tanggal_konsul",
                                    "status",
                                    "metode"
                                ],
                                "properties": {
                                    "id_pasien": {
                                        "type": "integer"
                                    },
                                    "id_dokter": {
                                        "type": "integer"
                                    },
                                    "tanggal_konsul": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "pending",
                                            "selesai",
                                            "batal"
                                        ]
                                    },
                                    "metode": {
                                        "type": "string",
                                        "enum": [
                                            "online",
                                            "offline"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Konsultasi created successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/konsultasis/{id}": {
            "get": {
                "tags": [
                    "Konsultasi"
                ],
                "summary": "Get a consultation by ID",
                "operationId": "4aac2ef233f198b3803481a0447237b1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Konsultasi retrieved successfully"
                    },
                    "404": {
                        "description": "Konsultasi not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Konsultasi"
                ],
                "summary": "Update a consultation by ID",
                "operationId": "9aab10d61eeefaace852b17d14862a5d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id_pasien",
                                    "id_dokter",
                                    "tanggal_konsul",
                                    "status",
                                    "metode"
                                ],
                                "properties": {
                                    "id_pasien": {
                                        "type": "integer"
                                    },
                                    "id_dokter": {
                                        "type": "integer"
                                    },
                                    "tanggal_konsul": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "pending",
                                            "selesai",
                                            "batal"
                                        ]
                                    },
                                    "metode": {
                                        "type": "string",
                                        "enum": [
                                            "online",
                                            "offline"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Konsultasi updated successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Konsultasi"
                ],
                "summary": "Delete consultation by ID",
                "operationId": "b14856ce8969e92290693722f82ee869",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Konsultasi deleted successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/konsultasis/detail": {
            "get": {
                "tags": [
                    "Konsultasi"
                ],
                "summary": "Menampilkan semua data konsultasi beserta relasi pasien dan dokter",
                "operationId": "3db3dd0131dfdb98e27748d81534cd73",
                "responses": {
                    "200": {
                        "description": "Konsultasi with relations retrieved successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Masukkan token dalam format: Bearer {token}"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}